requirements.txt
asgiref==3.11.1
Django>=5.0,<6.0
djangorestframework==3.17.0
psycopg2-binary==2.9.11
python-dotenv==1.2.2
sqlparse==0.5.5
redis==5.0.1
django-redis==5.4.0
celery==5.3.6
django-filter==24.2
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
pytest==8.2.0
pytest-django==4.8.0
django-debug-toolbar==4.4.6